requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
streamlit-drawable-canvas